MLCOMPUTE-1163 | handle space separated Spark args
