+++ paasta_tools/utils.py
@@ -2777,11 +2777,6 @@ class SystemPaastaConfig:
     def get_tron_k8s_use_suffixed_log_streams_k8s(self) -> bool:
         return self.config_dict.get("tron_use_suffixed_log_streams", False)
 
-    def get_spark_ui_port(self) -> int:
-        # 33000 was picked arbitrarily (it was the base port when we used to
-        # randomly reserve port numbers)
-        return self.config_dict.get("spark_ui_port", 33000)
-
     def get_spark_driver_port(self) -> int:
         # default value is an arbitrary value
         return self.config_dict.get("spark_driver_port", 33001)
